Fix int type warning

+++ src/libsrcml/srcml_unit.cpp
@@ -717,7 +717,7 @@ static int srcml_unit_unparse_internal(struct srcml_unit* unit, std::function<xm
         neol.reserve(unit->src->size());
 
         const std::string& src = *unit->src;
-        for (int i = 0; i < src.size(); ++i) {
+        for (size_t i = 0; i < src.size(); ++i) {
             if (src[i] != '\n') {
                 neol += src[i];
             }
